refactor(fleet-change): extract target planning
